[tests] split platform and backend tests out
This way we can avoid conflicts when merging into full.
