+++ tests/libvirt/mock_libvirt.cpp
