+++ tests/libvirt/mock_libvirt.h
