+++ tests/qemu/mock_dnsmasq_server.cpp
