+++ tests/qemu/mock_qemu_system.cpp
