+++ tests/qemu/mock_qemuimg.cpp
